requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.1.4
django-ninja==0.20.0
kombu==5.2.4
prompt-toolkit==3.0.36
psycopg2-binary==2.9.5
pydantic==1.10.2
PyMySQL==1.0.2
python-decouple==3.6
pytz==2022.7
redis==4.4.0
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.4.0
vine==5.0.0
wcwidth==0.2.5
